week-2 backend & frontend
