+++ client/src/App.jsx
@@ -1,4 +1,4 @@
-import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
+import { BrowserRouter as Router, Routes, Route, useLocation } from "react-router-dom";
 import Navbar from "./components/Navbar";
 import Footer from "./components/Footer";
 import Home from "./pages/LandingPage/Home";
@@ -12,27 +12,36 @@ import NotFound from './pages/NotFound.jsx';
 import { AuthProvider } from './context/AuthContext.jsx';
 import { Toaster } from "react-hot-toast";
 
+function Layout({ children }) {
+  const location = useLocation();
+  const hideLayout = location.pathname === "/chat" || location.pathname.startsWith("/chat/");
+  return (
+    <div className="min-h-screen flex flex-col bg-gray-50">
+      {!hideLayout && <Navbar />}
+      <Toaster position="top-center" />
+      <main className="flex-grow">{children}</main>
+      {!hideLayout && <Footer />}
+    </div>
+  );
+}
+
 export default function App() {
   return (
     <AuthProvider>
       <Router>
-        <div className="min-h-screen flex flex-col bg-gray-50">
-          <Navbar />
-          <Toaster position="top-center" /> 
-          <main className="flex-grow">
-            <Routes>
-              <Route path="/" element={<Home />} />
-              <Route path="/upload" element={<Upload />} />
-              <Route path="/chatbot" element={<ChatbotPage />} />
-              <Route path="/login" element={<Login />} />
-              <Route path= "/quiz" element={ <QuizPage /> }/>
-              <Route path="/signup" element={<Signup />} /> 
-              <Route path="/reset-password" element={<ResetPassword />} />
-              <Route path="*" element={<NotFound />}/>
-            </Routes>
-          </main>
-          <Footer />
-        </div>
+        <Layout>
+          <Routes>
+            <Route path="/" element={<Home />} />
+            <Route path="/upload" element={<Upload />} />
+            <Route path="/chat" element={<ChatbotPage />} />
+            <Route path="/chat/:id" element={<ChatbotPage />} />
+            <Route path="/login" element={<Login />} />
+            <Route path="/quiz" element={<QuizPage />} />
+            <Route path="/signup" element={<Signup />} />
+            <Route path="/reset-password" element={<ResetPassword />} />
+            <Route path="*" element={<NotFound />} />
+          </Routes>
+        </Layout>
       </Router>
     </AuthProvider>
   );
